Extract User::SpreadableAlerts
These methods don't interact with anything else inside User and all
relate to some incidental behaviour so are ripe for extraction.
