+++ app/models/user.rb
@@ -46,6 +46,7 @@ class User < ApplicationRecord
   include User::LoginToken
   include User::OneTimePassword
   include User::Slug
+  include User::SpreadableAlerts
   include User::Survey
   include Rails.application.routes.url_helpers
   include LinkToHelper
@@ -180,8 +181,6 @@ class User < ApplicationRecord
 
   validate :email_and_name_are_valid
 
-  after_initialize :set_defaults
-
   after_update :update_pro_account
   after_update :reindex_referencing_models, :invalidate_cached_pages,
                unless: :no_xapian_reindex
@@ -271,29 +270,6 @@ class User < ApplicationRecord
   # if they follow a /c/ redirect link belonging to another user?
   def self.stay_logged_in_on_redirect?(user)
     user&.is_admin?
-  end
-
-  # Used for default values of last_daily_track_email
-  def self.random_time_in_last_day
-    earliest_time = Time.zone.now - 1.day
-    latest_time = Time.zone.now
-    earliest_time + rand(latest_time - earliest_time).seconds
-  end
-
-  # Alters last_daily_track_email for every user, so alerts will be sent
-  # spread out fairly evenly throughout the day, balancing load on the
-  # server. This is intended to be called by hand from the Ruby console.  It
-  # will mean quite a few users may get more than one email alert the day you
-  # do it, so have a care and run it rarely.
-  #
-  # This SQL statement is useful for seeing how spread out users are at the moment:
-  # select extract(hour from last_daily_track_email) as h, count(*) from users group by extract(hour from last_daily_track_email) order by h;
-  def self.spread_alert_times_across_day
-    find_each do |user|
-      user.update!(last_daily_track_email: User.random_time_in_last_day)
-    end
-
-    nil # so doesn't print all users on console
   end
 
   def self.record_bounce_for_email(email, message)
@@ -697,19 +673,6 @@ class User < ApplicationRecord
   end
 
   private
-
-  def set_defaults
-    return unless new_record?
-
-    # make alert emails go out at a random time for each new user, so
-    # overall they are spread out throughout the day.
-    self.last_daily_track_email = self.class.random_time_in_last_day
-
-    # Make daily summary emails go out at a random time for each new user
-    # too, if it's not already set
-    self.daily_summary_hour ||= self.class.random_time_in_last_day.hour
-    self.daily_summary_minute ||= self.class.random_time_in_last_day.min
-  end
 
   def email_and_name_are_valid
     if email != "" && !MySociety::Validate.is_valid_email(email)
